project/requirements.txt
pyzmq == 19.0.2
python-consul